created two new routes
