+++ server/controllers/admin/userController.js
@@ -13,31 +13,6 @@ const JwtStrategy = passportJWT.Strategy
 
 const userController = {
 
-  // getCurrentUser
-  getCurrentUser: (req, res) => {
-
-    let TokenArray = req.headers.authorization.split(" ");
-    let authorization = TokenArray[1]
-
-    jwt.verify(authorization, process.env.JWT_SECRET, (err, authorizedData) => {
-
-      User.findOne({
-        where: {
-          id: authorizedData.id
-        }
-      }).then(user => {
-        return res.json({
-          id: user.id,
-          account: user.account,
-          phone: user.phone,
-          name: user.name,
-          email: user.email,
-          isAdmin: user.isAdmin
-        })
-      })
-    })
-  },
-
   // 登入
   signIn: (req, res) => {
